chacge cords type to float

+++ src/config.py
@@ -91,9 +91,9 @@ def config_parser():
 
     parser.add_argument("-f", "--fff", help="a dummy argument to fool ipython", default="1")
     
-    parser.add_argument("--x", type=int, required=False, help="x for generation")
-    parser.add_argument("--y", type=int, required=False, help="y for generation")
-    parser.add_argument("--z", type=int, required=False, help="z for generation")
+    parser.add_argument("--x", type=float, required=False, help="x for generation")
+    parser.add_argument("--y", type=float, required=False, help="y for generation")
+    parser.add_argument("--z", type=float, required=False, help="z for generation")
 
     args = parser.parse_args()
     return args
